requirements.txt
pinecone-client==2.2.2
openai==0.28.0
numpy==1.23.0
langchain==0.0.281
scikit-learn==1.3.0
streamlit==1.26.0
tiktoken==0.4.0